+++ src/packages/piral-cli/src/injectors/pilet.ts
@@ -76,6 +76,9 @@ export default class PiletInjector implements KrasInjector {
       injector: { name: this.name },
       headers: {
         'content-type': type,
+        'cache-control': 'no-cache, no-store, must-revalidate',
+        pragma: 'no-cache',
+        expires: 0,
       },
       status: { code: 200 },
       url,