+++ src/packages/piral-cli/src/common/bundler.ts
@@ -1,5 +1,7 @@
-import { resolve } from 'path';
-import { removeDirectory, checkExists } from './io';
+import { ParcelBundle } from 'parcel-bundler';
+import { transformFileAsync } from '@babel/core';
+import { resolve, dirname, basename } from 'path';
+import { removeDirectory, checkExists, readJson, writeText, writeJson } from './io';
 
 const bundleWithCodegen = require('parcel-plugin-codegen');
 
@@ -14,4 +16,64 @@ export async function clearCache(root: string) {
   if (exists) {
     await removeDirectory(cacheDir);
   }
+}
+
+interface BundleSource {
+  src: string;
+  map: string;
+}
+
+function gatherJsBundles(bundle: ParcelBundle, gatheredBundles: Array<BundleSource> = []) {
+  if (bundle.type === 'js') {
+    let map = undefined;
+
+    for (const childBundle of bundle.childBundles) {
+      if (childBundle.name.endsWith('.js.map')) {
+        map = childBundle.name;
+        break;
+      }
+    }
+
+    gatheredBundles.push({
+      src: bundle.name,
+      map,
+    });
+  }
+
+  for (const childBundle of bundle.childBundles) {
+    gatherJsBundles(childBundle, gatheredBundles);
+  }
+
+  return gatheredBundles;
+}
+
+/**
+ * Changes the files from the generated bundle to *really*
+ * follow the desired preset environment - which includes also
+ * packages coming from node_modules.
+ * @param mainBundle The main bundle coming from Parcel.
+ * @param rootDir The project's root dir.
+ */
+export function postTransform(mainBundle: ParcelBundle, rootDir: string) {
+  const bundles = gatherJsBundles(mainBundle);
+
+  return Promise.all(
+    bundles.map(async bundle => {
+      const inputSourceMap = bundle.map && (await readJson(dirname(bundle.map), basename(bundle.map)));
+
+      const { code, map } = await transformFileAsync(bundle.src, {
+        presets: [['@babel/preset-env']],
+        sourceMaps: true,
+        minified: true,
+        inputSourceMap,
+        cwd: rootDir,
+      });
+
+      await writeText(dirname(bundle.src), basename(bundle.src), code);
+
+      if (bundle.map) {
+        await writeJson(dirname(bundle.map), basename(bundle.map), map);
+      }
+    }),
+  );
 }