+++ src/packages/piral-core/src/state/createActions.test.ts
@@ -1,4 +1,4 @@
-import { createActions } from './createActions';
+import { createActions, includeActions } from './createActions';
 import { createListener } from '../utils';
 
 jest.mock('../actions', () => ({
@@ -30,5 +30,18 @@ describe('Create Actions Module', () => {
     const actions2 = createActions('bdf' as any, events);
     expect((actions1 as any).a()).toEqual('abc');
     expect((actions2 as any).a()).toEqual('bdf');
+  });
+
+  it('includeActions patches the initial creator', () => {
+    const events = createListener(undefined);
+    const actions = createActions('abc' as any, events);
+    includeActions(actions, {
+      c(arg) { return arg; },
+      d() { return 'bar'; },
+    } as any);
+    expect((actions as any).a()).toEqual('abc');
+    expect((actions as any).b()).toEqual('foo');
+    expect((actions as any).c()).toEqual('abc');
+    expect((actions as any).d()).toEqual('bar');
   });
 });