+++ src/packages/piral-cli/src/apps/build-piral.ts
@@ -21,6 +21,10 @@ import {
   findPackageVersion,
   coreExternals,
   combineApiDeclarations,
+  cliVersion,
+  postTransform,
+  logInfo,
+  ParcelConfig,
 } from '../common';
 
 interface Destination {
@@ -50,11 +54,10 @@ async function bundleFiles(
   target: string,
   dependencies: Array<string>,
   entryFiles: string,
-  detailedReport: boolean,
-  publicUrl: string,
-  logLevel: 1 | 2 | 3,
   dest: Destination,
   subdir: string,
+  config: ParcelConfig,
+  transformRoot?: string,
 ) {
   const outDir = join(dest.outDir, subdir);
 
@@ -71,22 +74,21 @@ async function bundleFiles(
   const bundler = new Bundler(
     entryFiles,
     extendConfig({
+      ...config,
       outDir,
       outFile: dest.outFile,
-      watch: false,
-      minify: true,
-      scopeHoist: false,
-      contentHash: true,
-      detailedReport,
-      publicUrl,
-      logLevel,
     }),
   );
 
   extendBundlerForPiral(bundler);
   extendBundlerWithPlugins(bundler);
 
-  await bundler.bundle();
+  const bundle = await bundler.bundle();
+
+  if (transformRoot && config.minify) {
+    await postTransform(bundle, transformRoot);
+  }
+
   return outDir;
 }
 
@@ -101,21 +103,31 @@ export type PiralBuildType = 'all' | 'release' | 'develop';
 export interface BuildPiralOptions {
   entry?: string;
   target?: string;
+  cacheDir?: string;
   publicUrl?: string;
+  minify?: boolean;
   detailedReport?: boolean;
   logLevel?: 1 | 2 | 3;
   fresh?: boolean;
   type?: PiralBuildType;
+  sourceMaps?: boolean;
+  contentHash?: boolean;
+  scopeHoist?: boolean;
 }
 
 export const buildPiralDefaults = {
   entry: './',
   target: './dist',
   publicUrl: '/',
+  cacheDir: '.cache',
   detailedReport: false,
   logLevel: 3 as const,
   fresh: false,
+  minify: true,
   type: 'all' as const,
+  sourceMaps: true,
+  contentHash: true,
+  scopeHoist: false,
 };
 
 export async function buildPiral(baseDir = process.cwd(), options: BuildPiralOptions = {}) {
@@ -125,6 +137,11 @@ export async function buildPiral(baseDir = process.cwd(), options: BuildPiralOpt
     publicUrl = buildPiralDefaults.publicUrl,
     detailedReport = buildPiralDefaults.detailedReport,
     logLevel = buildPiralDefaults.logLevel,
+    cacheDir = buildPiralDefaults.cacheDir,
+    minify = buildPiralDefaults.minify,
+    sourceMaps = buildPiralDefaults.sourceMaps,
+    contentHash = buildPiralDefaults.contentHash,
+    scopeHoist = buildPiralDefaults.scopeHoist,
     fresh = buildPiralDefaults.fresh,
     type = buildPiralDefaults.type,
   } = options;
@@ -135,26 +152,30 @@ export async function buildPiral(baseDir = process.cwd(), options: BuildPiralOpt
   const dest = getDestination(entryFiles, resolve(baseDir, target));
 
   if (fresh) {
-    await clearCache(root);
+    await clearCache(root, cacheDir);
   }
 
   await removeDirectory(dest.outDir);
 
   // everything except release -> build develop
   if (type !== 'release') {
+    logInfo('Starting build ...');
+
+    // we'll need this info for later
+    const originalPackageJson = resolve(root, 'package.json');
+    const { files: originalFiles = [] } = require(originalPackageJson);
     const appDir = 'app';
-    const outDir = await bundleFiles(
-      name,
-      true,
-      targetDir,
-      externals,
-      entryFiles,
+    const outDir = await bundleFiles(name, true, targetDir, externals, entryFiles, dest, join('develop', appDir), {
+      cacheDir,
+      watch: false,
+      sourceMaps,
+      contentHash,
+      minify,
+      scopeHoist,
       detailedReport,
       publicUrl,
       logLevel,
-      dest,
-      join('develop', appDir),
-    );
+    });
     const allExternals = [...externals, ...coreExternals];
     const externalPackages = await Promise.all(
       allExternals.map(async name => ({
@@ -185,6 +206,10 @@ export async function buildPiral(baseDir = process.cwd(), options: BuildPiralOpt
         ...pilets,
         files,
       },
+      piralCLI: {
+        version: cliVersion,
+        generated: true,
+      },
       main: `${appDir}/index.js`,
       typings: `${appDir}/index.d.ts`,
       app: `${appDir}/index.html`,
@@ -196,28 +221,50 @@ export async function buildPiral(baseDir = process.cwd(), options: BuildPiralOpt
       },
     });
     await createDirectory(filesDir);
+    // for scaffolding we need to keep the files also available in the new package
     await copyScaffoldingFiles(root, filesDir, pilets.files);
+    // we just want to make sure that "files" mentioned in the original package.json are respected in the package
+    await copyScaffoldingFiles(root, rootDir, originalFiles);
+    // actually including this one hints that the app shell should have been included - which is forbidden
     await createFileIfNotExists(outDir, 'index.js', 'throw new Error("This file should not be included anywhere.");');
     await generateDeclaration(outDir, root, name, dependencies.std);
     await createPackage(rootDir);
     await Promise.all([removeDirectory(outDir), removeDirectory(filesDir), remove(resolve(rootDir, 'package.json'))]);
-    logDone(`Development package available in "${rootDir}".\n`);
+
+    logDone(`Development package available in "${rootDir}".`);
+  }
+
+  if (type === 'all') {
+    // Just have some space between the two builds
+    logInfo('\n\n\n\n\n\n');
   }
 
   // everything except develop -> build release
   if (type !== 'develop') {
+    logInfo('Starting build ...');
+
     const outDir = await bundleFiles(
       name,
       false,
       targetDir,
       externals,
       entryFiles,
-      detailedReport,
-      publicUrl,
-      logLevel,
       dest,
       'release',
+      {
+        cacheDir,
+        watch: false,
+        sourceMaps,
+        contentHash,
+        minify,
+        scopeHoist,
+        detailedReport,
+        publicUrl,
+        logLevel,
+      },
+      root,
     );
-    logDone(`Files for publication available in "${outDir}".\n`);
+
+    logDone(`Files for publication available in "${outDir}".`);
   }
 }