+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Fixed a bug in `piral-urql` when options are not set
 * Introduced a sanity check for debugging pilets (#115)
+* Removed potential caching from `pilet debug` (#117)
 
 ## 0.9.2 (December 11, 2019)