+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Piral Changelog
 
+## 0.9.3 (December 17, 2019)
+
+* Fixed a bug in `piral-urql` when options are not set
+* Introduced a sanity check for debugging pilets (#115)
+* Removed potential caching from `pilet debug` (#117)
+* Copy the files as specified when creating the dev package
+* Added final bundle transformation according to the specified preset (#114)
+* Provided ability to set initial custom actions
+* Exposed the full Parcel CLI spectrum in the `piral-cli` (#119)
+
 ## 0.9.2 (December 11, 2019)
 
 * Fixed not found error when copying files for scaffold (#106)